add ENS Records fetching to Store Provider
