+++ src/features/Ens/EnsDashboard.tsx
@@ -1,18 +1,14 @@
-import React, { useContext, useState } from 'react';
+import React, { useContext } from 'react';
 
 import { Heading } from '@mycrypto/ui';
 import styled from 'styled-components';
 
 import { DashboardPanel, PoweredByText } from '@components';
 import { StoreContext } from '@services';
-import EnsService from '@services/ApiService/Ens/EnsService.ts';
-import { isEthereumAccount } from '@services/Store/Account/helpers';
 import { BREAK_POINTS, SPACING } from '@theme';
 import { translateRaw } from '@translations';
-import { useEffectOnce, usePromise } from '@vendor/react-use';
 
 import { EnsTable } from './EnsTable';
-import { DomainNameRecord } from './types';
 
 const DashboardWrapper = styled.div`
   width: 100%;
@@ -34,37 +30,8 @@ const StyledLayout = styled.div`
   }
 `;
 
-export interface MyDomainsData {
-  records: DomainNameRecord[];
-  isFetched: boolean;
-}
-
-const defaultData: MyDomainsData = {
-  records: [] as DomainNameRecord[],
-  isFetched: false
-};
-
 export default function EnsDashboard() {
-  const { accounts } = useContext(StoreContext);
-  const [ensOwnershipRecords, setEnsOwnershipRecords] = useState(defaultData);
-  const mounted = usePromise();
-
-  // Only use the accounts on the Ethereum mainnet network
-  const accountsEthereumNetwork = accounts.filter(isEthereumAccount);
-
-  useEffectOnce(() => {
-    (async () => {
-      const ownershipRecords: MyDomainsData = await mounted(
-        EnsService.fetchOwnershipRecords(accountsEthereumNetwork).then(
-          (data: DomainNameRecord[]) => ({
-            records: data,
-            isFetched: true
-          })
-        )
-      );
-      setEnsOwnershipRecords(ownershipRecords);
-    })();
-  });
+  const { ensOwnershipRecords } = useContext(StoreContext);
 
   return (
     <StyledLayout>
@@ -76,10 +43,7 @@ export default function EnsDashboard() {
           heading={translateRaw('ENS_MY_DOMAINS_TABLE_HEADER')}
           headingRight={<PoweredByText provider="ENS" />}
         >
-          <EnsTable
-            records={ensOwnershipRecords.records}
-            isFetched={ensOwnershipRecords.isFetched}
-          />
+          <EnsTable records={ensOwnershipRecords} />
         </DashboardPanel>
       </DashboardWrapper>
     </StyledLayout>
