+++ src/features/Ens/EnsTable.tsx
@@ -2,35 +2,19 @@ import React from 'react';
 
 import styled from 'styled-components';
 
-import { Spinner } from '@components';
+import { DomainNameRecord } from '@types';
 
-import { MyDomainsData } from './EnsDashboard';
 import MyDomains from './MyDomains';
 import NoDomains from './NoEnsDomains';
-
-const SpinnerContainer = styled.div`
-  height: 400px;
-  display: float;
-  align-items: center;
-  justify-content: center;
-`;
 
 const EnsTableContainer = styled.div`
   max-height: 650px;
   overflow: auto;
 `;
 
-export const EnsTable = ({ records, isFetched }: MyDomainsData) => (
+export const EnsTable = ({ records }: { records: DomainNameRecord[] }) => (
   <EnsTableContainer>
-    {records.length === 0 &&
-      (isFetched ? (
-        <NoDomains />
-      ) : (
-        <SpinnerContainer>
-          <Spinner size={'x3'} />
-        </SpinnerContainer>
-      ))}
-    {records.length !== 0 && <MyDomains domainOwnershipRecords={records} />}
+    {records.length === 0 ? <NoDomains /> : <MyDomains domainOwnershipRecords={records} />}
   </EnsTableContainer>
 );
