+++ src/features/Ens/__tests__/EnsDashboard.test.tsx
@@ -3,27 +3,20 @@ import React from 'react';
 import { simpleRender } from 'test-utils';
 
 import { translateRaw } from '@translations';
+import { DomainNameRecord } from '@types';
 
 import { EnsTable } from '../EnsTable';
-import { DomainNameRecord } from '../types';
 
 /* Test components */
 describe('ENSTable', () => {
-  const component = (records: DomainNameRecord[], isFetched: boolean) => (
-    <EnsTable records={records} isFetched={isFetched} />
-  );
+  const component = (records: DomainNameRecord[]) => <EnsTable records={records} />;
 
-  const renderComponent = (records: DomainNameRecord[], isFetched: boolean) => {
-    return simpleRender(component(records, isFetched));
+  const renderComponent = (records: DomainNameRecord[]) => {
+    return simpleRender(component(records));
   };
 
-  test('Can render loading spinner state of component', () => {
-    const { container } = renderComponent([], false);
-    expect(container.querySelector('svg')).toBeDefined();
-  });
-
   test('Can render no-domains state of component', () => {
-    const { getByText } = renderComponent([], true);
+    const { getByText } = renderComponent([]);
     const selector = translateRaw('ENS_DOMAINS_NO_DOMAINS').trim();
     expect(getByText(selector)).toBeInTheDocument();
   });
@@ -38,7 +31,7 @@ describe('ENSTable', () => {
         readableDomainName: 'donate.mycryptoid.eth'
       }
     ];
-    const { getByText } = renderComponent(testRecords, true);
+    const { getByText } = renderComponent(testRecords);
     const selector = translateRaw('ENS_MY_DOMAINS_TABLE_EXPIRES_HEADER').trim();
     expect(getByText(selector)).toBeInTheDocument(); // Expect to see the Network selection step
   });
