+++ src/features/Ens/types.ts
@@ -1,3 +1,5 @@
+import { DomainNameRecord } from '@types';
+
 export interface MyDomainsTableProps {
   domainName: string;
   owner: string;
@@ -6,31 +8,4 @@ export interface MyDomainsTableProps {
 
 export interface MyDomainsProps {
   domainOwnershipRecords: DomainNameRecord[];
-}
-
-export interface DomainEntry {
-  expiryDate: number;
-  domain: DomainChild;
-}
-
-export interface DomainChild {
-  isMigrated: boolean;
-  labelName: string;
-  labelhash: string;
-  name: string;
-  parent: DomainParent;
-  __typename: string;
-}
-
-export interface DomainParent {
-  name: string;
-  __typename: string;
-}
-
-export interface DomainNameRecord {
-  owner: string;
-  ownerLabel: string;
-  domainName: string;
-  expiryDate: string;
-  readableDomainName: string;
 }
