+++ src/types/ens.ts
@@ -1,0 +1,26 @@
+export interface DomainEntry {
+  expiryDate: number;
+  domain: DomainChild;
+}
+
+export interface DomainChild {
+  isMigrated: boolean;
+  labelName: string;
+  labelhash: string;
+  name: string;
+  parent: DomainParent;
+  __typename: string;
+}
+
+export interface DomainParent {
+  name: string;
+  __typename: string;
+}
+
+export interface DomainNameRecord {
+  owner: string;
+  ownerLabel: string;
+  domainName: string;
+  expiryDate: string;
+  readableDomainName: string;
+}
